LayoutTests/fast/canvas/drawImage-with-negative-source-destination-expected.txt

Series of tests to ensure correct behaviour on negative source/destination of a HTMLCanvasElement

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[4] is 0
PASS imgdata[5] is 255
PASS imgdata[6] is 0
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[4] is 0
PASS imgdata[5] is 0
PASS imgdata[6] is 0
PASS successfullyParsed is true

TEST COMPLETE
